fix: windows compatibility
- fix file write error
- fix datetime to timestamp error
- fix read file line by line then json.parse error
- test, add windows test actions
- build, remove useless test packages
- test, fix windows platform not work with unicode characters in codecov config
