+++ setup.py
@@ -42,10 +42,8 @@ setup(
         'Operating System :: OS Independent',
         'Programming Language :: Python',
         'Programming Language :: Python :: 2',
-        'Programming Language :: Python :: 2.6',
         'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.3',
         'Programming Language :: Python :: 3.4',
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
@@ -66,8 +64,6 @@ setup(
             'pytest',
             'pytest-cov',
             'freezegun',
-            'scrutinizer-ocular',
-            'codecov'
         ]
     },
